Separate processing of model data and the rest
